Remove UserEntryExt file and move UserEntryPresentationHelper.kt in utils folder
