+++ core/src/main/res/drawable/ic_food.xml
