+++ core/src/main/res/drawable/ic_generic_icon.xml
