+++ core/src/main/res/drawable/ic_sms.xml
